refactor(whale_lair): make code easier to read
